Pulled out sharable code into utils script - created utils script - pulled out code that is similar - added JSDocs to explain utils
